requirements.txt
# requirements.txt
aiogram==3.13.1
aiohttp==3.10.11
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
httpx==0.28.1
yookassa==3.4.0
python-dotenv==1.0.1